Up to date restructuring
