--- app/flashcards/test card with html.csv
@@ -1,1 +1,0 @@
-pero poco después el thestral <strong>desplegó las alas</strong> con un contundente movimiento que casi derribó al chico,spread its wings,,,,,,Spanish,English
