--- app/flashcards/test html.csv
@@ -1,5 +1,0 @@
-"<!DOCTYPE HTML PUBLIC ""-//W3C//DTD HTML 4.0//EN"" ""http://www.w3.org/TR/REC-html40/strict.dtd"">
-<html><head><meta name=""qrichtext"" content=""1"" /><style type=""text/css"">
-p, li { white-space: pre-wrap; }
-</style></head><body style="" font-family:'MS Shell Dlg 2'; font-size:8pt; font-weight:400; font-style:normal;"">
-<p style="" margin-top:0px; margin-bottom:0px; margin-left:0px; margin-right:0px; -qt-block-indent:0; text-indent:0px;"">Hola mi <span style="" font-weight:600;"">mejor</span> amigo</p></body></html>",best,,,,,,Spanish,English
