+++ app/ui/sentence_bin.py
